refactor(visualize_dataset_distribution.py, utils.py): move get_all_categories to imported file
